MDS-4625 Changed location of tsfs in core to under Permits & Approvals tab
