+++ services/core-web/src/components/mine/Tailings/MineReportTailingsInfo.js
@@ -1,0 +1,6 @@
+import React from "react";
+import MineTailingsInfoTabs from "./MineTailingsInfoTabs";
+
+const MineReportTailingsInfo = () => <MineTailingsInfoTabs enabledTabs={["reports", "map"]} />;
+
+export default MineReportTailingsInfo;
