+++ services/core-web/src/routes/MineDashboardRoutes.js
@@ -54,6 +54,11 @@ const MineDashboardRoutes = () => (
       path={routes.MINE_PERMIT_REQUIRED_REPORTS.route}
       component={routes.MINE_PERMIT_REQUIRED_REPORTS.component}
     />
+    <Route
+      exact
+      path={routes.MINE_TAILINGS_REPORTS.route}
+      component={routes.MINE_TAILINGS_REPORTS.component}
+    />
     <Route exact path={routes.MINE_TAILINGS.route} component={routes.MINE_TAILINGS.component} />
     <Route exact path={routes.MINE_DOCUMENTS.route} component={routes.MINE_DOCUMENTS.component} />
     <Route
